+++ cdf/domain/latest/__init__.py
@@ -2,5 +2,5 @@ from .tracking import *
 from .event import *
 from .match import *
 from .meta import *
-from .skeletal import *
+from .landmark import *
 from .video import *